requirements.txt
aiofiles==0.6.0
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
blinker==1.4
chardet==3.0.4
click==7.1.2
discord.py==1.5.1
emoji==0.6.0
Flask==1.1.2
h11==0.11.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
priority==1.3.0
pycodestyle==2.6.0
python-dotenv==0.14.0
Quart==0.13.1
soupsieve==2.0.1
SQLAlchemy==1.3.20
toml==0.10.1
Werkzeug==1.0.1
wsproto==0.15.0
yarl==1.5.1
